fix: use eslint to format ts,js,json,md and yml, goodbye prettier
prettier is great but introduces additional configuration(config, ignores and scripts), and packages. Eslint is pretty much capable of doing eveything and gives you much control over the opinionated nature of prettier.
Prettier describes itself to be "an opinionated code formatter".
Opinionated essentially means it’s not for everyone. Prettier makes a lot of hard-coded decisions to provide a minimal configuration interface. That makes it very easy to use (it’s excellent!) and the code consistent across projects. However, on the other hand, it also means you are losing the ability to have fine-grained tweaks to how your code should look like.